Add tests for Template ignore_undefined_variables parameter
Test if it is possible to ignore images variables in Templates. Test if it is
possible to ignore if and for loops. Test if it is not possible to ignore dotted
variables like py3o.var.value. In Genshi it is currently not possible to ignore
such variables (e.g. Django templates would have ignored them too).