[REF]main.py: fix generated VALUE formula in `__prepare_calc_formulas` function
If the Genshi expressions generates a number value (float or integer),
the generated formula will be `=VALUE("1.1")` which is problematic if the
user machine is configured to use a decimal separator different than the `.`.
With this fix the generated formula will be `=VALUE(1.1)` (without the quotes)
which is correctly handled with by the `.ods` format.